角色权限修改

See merge request industrial/smartindustry!247

+++ authoritymanage/src/main/java/com/smartindustry/authority/service/impl/RoleServiceImpl.java
@@ -177,8 +177,10 @@ public class RoleServiceImpl implements IRoleService {
             // 先根据角色id找出所有符合用户角色的用户id
             List<Long> uids = userMapper.queryUser(dto.getRid());
             // 再更新用户权限中间表
-            mUserAuthorityMapper.deleteByUserId(uids);
-            mUserAuthorityMapper.insertBatch(uids,dto.getPerms());
+            if(!uids.isEmpty()){
+                mUserAuthorityMapper.deleteByUserId(uids);
+                mUserAuthorityMapper.insertBatch(uids,dto.getPerms());
+            }
         }
         return ResultVO.ok();
     }
